requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
